handlers: add learning handlers
