+++ services/api/app/diabetes/handlers/common_handlers.py
@@ -5,8 +5,8 @@ from __future__ import annotations
 from telegram import Update
 from telegram.ext import ContextTypes
 
-from services.api.app.config import settings
 from services.api.app.diabetes.utils.ui import menu_keyboard
+from .learning_handlers import learn_command
 
 
 async def menu_command(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
@@ -73,18 +73,6 @@ async def smart_input_help(update: Update, context: ContextTypes.DEFAULT_TYPE) -
     message = update.message
     if message:
         await message.reply_text(text, parse_mode="Markdown")
-
-
-async def learn_command(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
-    """Provide learning resources if the feature is enabled."""
-
-    message = update.message
-    if message is None:
-        return
-    if not settings.learning_enabled:
-        await message.reply_text("🚫 Обучение недоступно.")
-        return
-    await message.reply_text("📘 Учебный режим активен.")
 
 
 __all__ = [
