+++ services/api/app/diabetes/handlers/registration.py
@@ -151,7 +151,7 @@ def register_handlers(
         photo_handlers,
         sugar_handlers,
         gpt_handlers,
-        learn_handlers,
+        learning_handlers,
         billing_handlers,
     )
 
@@ -167,7 +167,7 @@ def register_handlers(
     app.add_handler(sos_handlers.sos_contact_conv)
     app.add_handler(CommandHandlerT("cancel", dose_calc.dose_cancel))
     app.add_handler(CommandHandlerT("help", help_command))
-    app.add_handler(CommandHandlerT("learn", learn_handlers.learn_command))
+    app.add_handler(CommandHandlerT("learn", learning_handlers.learn_command))
     app.add_handler(CommandHandlerT("gpt", gpt_handlers.chat_with_gpt))
     app.add_handler(CommandHandlerT("trial", billing_handlers.trial_command))
     app.add_handler(CommandHandlerT("upgrade", billing_handlers.upgrade_command))
